Fixed fuel bug
What happened was that that this line was looking for a signed 32-bit value but was given an unsigned 32-bit value. 
Meaning the R.U. you get from the Infinite R.U. cheat and the R.U. given to you by the Chmmr looked like negative values to this line in the code. 
Making it impossible to refill your fuel tanks.